requirements.txt
torch
torchaudio
wandb
tqdm
numpy
pyyaml
kaggle
torchinfo